Move compiled tests to same module
